add CHANGELOG
this makes it easier for consumers of the project to understand what has
changed in the project with each new release.
the idea is not to list a 1:1 copy of the commits done in the
repository, instead this lists the notable changes as seen from a
consumer point of view (e.g. internal refactoring isn't relevant here).